requirements.txt
fastapi
uvicorn[standard]
python-dotenv
aiosmtplib
jinja2
python-multipart
